约束调整

+++ YLT_Tangram/Classes/VC/YLT_TangramVC+Delegate.m
@@ -31,7 +31,7 @@
         return size;
     }
     
-    return CGSizeMake(item.layoutWidth.floatValue, item.layoutHeight.floatValue);
+    return CGSizeMake(item.layoutWidth, item.layoutHeight);
 }
 
 - (CGFloat)collectionView:(UICollectionView *)collectionView layout:(UICollectionViewLayout *)collectionViewLayout minimumLineSpacingForSectionAtIndex:(NSInteger)section {
